+++ PluginCore/Plugin.cs
@@ -1,20 +1,14 @@
 ﻿using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Logging;
-using HarmonyLib;
 using ShipColors.ConfigManager;
-using ShipColors.Customizer;
 using ShipColors.Events;
-using Steamworks.Data;
-using System.Linq;
-using System.Reflection;
-using UnityEngine;
 
 
 namespace ShipColors
 {
     [BepInPlugin("darmuh.ShipColors", "ShipColors", (PluginInfo.PLUGIN_VERSION))]
-    [BepInDependency("darmuh.OpenLib", "0.2.13")]
+    [BepInDependency("darmuh.OpenLib", "0.2.14")]
 
 
     public class Plugin : BaseUnityPlugin
@@ -24,7 +18,7 @@ namespace ShipColors
         {
             public const string PLUGIN_GUID = "darmuh.ShipColors";
             public const string PLUGIN_NAME = "ShipColors";
-            public const string PLUGIN_VERSION = "0.3.2";
+            public const string PLUGIN_VERSION = "0.3.3";
         }
 
         internal static ManualLogSource Log;